% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geva-package.R
\docType{package}
\name{geva-package}
\alias{geva}
\alias{geva-package}
\title{geva: Gene Expression Variation Analysis (GEVA)}
\description{
Statistic methods to evaluate variations of differential expression (DE)
  between multiple biological conditions.
  It takes into account the fold-changes and p-values from previous differential expression (DE)
  results that use large-scale data (*e.g.*, microarray and RNA-seq) and evaluates which genes
  would react in response to the distinct experiments.
  This evaluation involves an unique pipeline of statistical methods, including weighted summarization,
  quantile detection, cluster analysis, and ANOVA tests, in order to classify a subset of relevant genes
  whose DE is similar or dependent to certain biological factors.

Statistic methods to evaluate variation between multiple biological conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sbcblab/geva}
}

}
\author{
\strong{Maintainer}: Itamar José Guimarães Nunes \email{nunesijg@gmail.com} (\href{https://orcid.org/0000-0002-6246-4658}{ORCID})

Other contributors:
\itemize{
  \item Murilo Zanini David \email{zanini.murilo@gmail.com} [contributor]
  \item Bruno César Feltes \email{bcfeltes@gmail.com} (\href{https://orcid.org/0000-0002-2825-8295}{ORCID}) [contributor]
  \item Marcio Dorn \email{mdorn@inf.ufrgs.br} (\href{https://orcid.org/0000-0001-8534-3480}{ORCID}) [contributor]
}

}
\keyword{internal}
