% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\docType{data}
\name{options.quantiles}
\alias{options.quantiles}
\alias{geva.quantiles}
\title{Returns a vector with the supported methods of quantiles separation}
\format{
An object of class \code{character} of length 5.
}
\usage{
options.quantiles

geva.quantiles(
  sv,
  nq.s = 3L,
  nq.v = 2L,
  quantile.method = options.quantiles,
  initial.thresholds = c(S = NA_real_, V = NA_real_),
  comb.score.fn = prod,
  ...
)
}
\arguments{
\item{sv}{SVTable}

\item{nq.s}{number of quantiles in S-axis}

\item{nq.v}{number of quantiles in V-axis}

\item{quantile.method}{method to detect the initial quantile thresholds. Ignored if initial.thresholds are set}

\item{initial.thresholds}{named numeric vector with the threshold used for the first quantile}

\item{comb.score.fn}{function to merge S and V scores into a single column. prod and mean are some examples}
}
\description{
Returns a vector with the supported methods of quantiles separation

Calculates the quantiles of a SVTable
}
\keyword{datasets}
