% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\name{geva.summarize}
\alias{geva.summarize}
\title{Summarizes the GEVAInput}
\usage{
geva.summarize(
  gevainput,
  summary.method = options.summary,
  variation.method = options.variation
)
}
\description{
Summarizes the GEVAInput
}
