% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVACluster.R
\docType{class}
\name{GEVACluster-class}
\alias{GEVACluster-class}
\title{GEVA Clustering Results}
\description{
The \code{GEVACluster} class inherits the grouping properties from \code{GEVAGroupSet}.
}
\section{Slots}{

\describe{
\item{\code{cluster.method}}{Method used to obtain the clustering results}
}}

