% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVAGroupedSummary.R
\docType{class}
\name{GEVAGroupedSummary-class}
\alias{GEVAGroupedSummary-class}
\title{GEVA Grouped Summary-Variation Table}
\description{
The \code{GEVAGroupedSummary} class inherits from \code{GEVASummary}.
}
\section{Slots}{

\describe{
\item{\code{groupsetlist}}{TypedList containing elements of GEVAGroupSet-class}
}}

