% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVAInput.R
\docType{class}
\name{GEVAInput-class}
\alias{GEVAInput-class}
\title{GEVA Input Data}
\description{
The \code{GEVAInput} class contains the initial data for \code{GEVA} usage.
It stores numeric matrices of \emph{logFC} values from differential expression comparison results. Options for calculations and summarizing are also included.
}
\section{Slots}{

\describe{
\item{\code{values}}{numeric matrix (m*n) of log-ratio values, usually \emph{logFC}}

\item{\code{weights}}{numeric matrix (m*n) of weighted values. If not defined, all weight values are equal to 1}

\item{\code{factors}}{factor representing the grouping of the n columns. If not defined, all factors are equal to NA}

\item{\code{ftable}}{data.frame with m rows containing attribute columns associated to the features (e.g. probes or genes)}

\item{\code{info}}{list containg additional information included in the input}
}}

