% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\docType{data}
\name{options.summary}
\alias{options.summary}
\title{Returns a vector with the supported methods of summarization}
\format{An object of class \code{character} of length 2.}
\usage{
options.summary
}
\description{
Returns a vector with the supported methods of summarization
}
\keyword{datasets}
