% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\docType{data}
\name{options.variation}
\alias{options.variation}
\title{Returns a vector with the supported methods of summarization}
\format{
An object of class \code{character} of length 3.
}
\usage{
options.variation
}
\description{
Returns a vector with the supported methods of summarization
}
\keyword{datasets}
