% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_SVAttribute.R
\docType{class}
\name{SVAttribute-class}
\alias{SVAttribute-class}
\title{Summary-Variation Attribute Field}
\description{
This simple S4 class stores two character slots representing attribute fields for summary and variation.
}
\section{Slots}{

\describe{
\item{\code{S}}{character or numeric of length one}

\item{\code{V}}{character or numeric of length one}
}}

