% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_SVTable.R
\docType{class}
\name{SVTable-class}
\alias{SVTable-class}
\title{Summary-Variation Table}
\description{
The \code{SVTable} class stores a \code{matrix} composed by two columns: \code{summary} and \code{variation}.
}
\section{Slots}{

\describe{
\item{\code{sv}}{matrix composed by two columns: summary and variation}
}}

