% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_TypedList.R
\docType{class}
\name{TypedList-class}
\alias{TypedList-class}
\title{Type-strict List (TypedList-class)}
\description{
List containing elements of the same class.
}
\section{Slots}{

\describe{
\item{\code{elem.class}}{character representing the class related to the elements}
}}

