% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geva-package.R
\docType{package}
\name{geva-package}
\alias{geva}
\alias{geva-package}
\title{geva: Gene Expression Variation Analysis}
\description{
Statistic methods to evaluate variation between multiple biological conditions.

Statistic methods to evaluate variation between multiple biological conditions.
}
\keyword{internal}
