% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{geva.quantiles}
\alias{geva.quantiles}
\title{Calculates the quantiles of a SVTable}
\usage{
geva.quantiles(
  sv,
  nq.s = 3L,
  nq.v = 2L,
  initial.thresholds = c(S = NA_real_, V = NA_real_),
  comb.score.fn = prod,
  ...
)
}
\arguments{
\item{sv}{SVTable}

\item{nq.s}{number of quantiles in S-axis}

\item{nq.v}{number of quantiles in V-axis}

\item{initial.thresholds}{named numeric vector with the threshold used for the first quantile}

\item{comb.score.fn}{function to merge S and V scores into a single column. prod and mean are some examples}
}
\description{
Calculates the quantiles of a SVTable
}
