% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarization.R
\docType{data}
\name{options.summary}
\alias{options.summary}
\alias{options.variation}
\alias{geva.summarize}
\title{Returns a vector with the supported methods of summarization}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
options.summary

options.variation

geva.summarize(
  gevainput,
  summary.method = options.summary,
  variation.method = options.variation
)
}
\description{
Returns a vector with the supported methods of summarization

Returns a vector with the supported methods of summarization

Summarizes the GEVAInput
}
\keyword{datasets}
