% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVAGroupSet.R
\docType{class}
\name{GEVAGroupSet-class}
\alias{GEVAGroupSet-class}
\title{GEVA Grouping Results}
\description{
The \code{GEVAGroupSet} class stores a factor that confers the grouping of summary/variation (SV) data.
}
\section{Slots}{

\describe{
\item{\code{grouping}}{factor used to group genes or probes}

\item{\code{scores}}{numeric vector with grouping scores}

\item{\code{ftable}}{data.frame with additional information related to the grouped features (genes or probes)}

\item{\code{centroids}}{SVTable describing the position of centroids}

\item{\code{offsets}}{SVTable with distances between the centroids and SV data}

\item{\code{info}}{list with additional information}
}}

