% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVAQuantiles.R, R/c_GEVAQuantilesAdjusted.R
\docType{class}
\name{GEVAQuantiles-class}
\alias{GEVAQuantiles-class}
\alias{GEVAQuantilesAdjusted-class}
\title{GEVA Quantiles Grouping Results}
\description{
The \code{GEVAQuantiles} class inherits the grouping properties from \code{GEVAGroupSet} and additionally stores the indexes associated to summary and variation.
}
\section{Slots}{

\describe{
\item{\code{svscores}}{SVTable with individual scores for summary and variation}

\item{\code{qareasizes}}{SVTable with nominal area sizes of the quantiles in terms of summary and variation}

\item{\code{qindexes}}{SVTable with indexes representing the magnitude index of summary and variation}

\item{\code{qcount}}{SVIntAttribute with number of quantiles for summary and variation}

\item{\code{qcutoff}}{SVNumAttribute with the initial quantile cut-offs, starting from zero}

\item{\code{grouprels}}{typed list of named factors representing external group relationships to the current quantiles}
}}

