% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVAQuantiles.R
\docType{class}
\name{GEVAQuantiles-class}
\alias{GEVAQuantiles-class}
\alias{show,GEVAQuantiles-method}
\alias{plot,GEVAQuantiles,SVTable-method}
\alias{[,GEVAQuantiles,ANY,ANY,ANY-method}
\alias{qindexes,GEVAQuantiles-method}
\alias{qareasizes,GEVAQuantiles-method}
\alias{qcount,GEVAQuantiles-method}
\alias{sv.scores,GEVAQuantiles-method}
\alias{cluster.method,GEVAQuantiles-method}
\alias{quantiles.method,GEVAQuantiles-method}
\alias{quantiles,GEVAQuantiles-method}
\alias{dim,GEVAQuantiles-method}
\alias{lines.GEVAQuantiles}
\alias{as.expression.GEVAQuantiles}
\alias{as.SVTable.GEVAQuantiles}
\title{GEVA Quantiles Grouping Results}
\usage{



























}
\description{
The \code{GEVAQuantiles} class represents the results of a quantile detection analysis. For each probe/gene, there is a assigned quantile among the \emph{g} defined quantiles.

This class inherits from \code{\linkS4class{GEVAGroupSet}} and is inherited by \code{\linkS4class{GEVAQuantilesAdjusted}}.
}
\section{Slots}{

\describe{
\item{\code{grouping}}{\code{factor} (\emph{m} elements, \emph{g} levels), quantile assignment for each gene/probe
\cr (Inherited from \code{\linkS4class{GEVAGroupSet}})}

\item{\code{scores}}{\code{numeric} vector (\emph{m} elements) with the assigned quantile scores for each gene/probe
\cr (Inherited from \code{\linkS4class{GEVAGroupSet}})}

\item{\code{ftable}}{\code{data.frame} (\emph{m} lines) with additional quantile assignment features
\cr (Inherited from \code{\linkS4class{GEVAGroupSet}})}

\item{\code{centroids}}{\code{numeric SVTable} (\emph{g} lines) with the S and V centroid coordinates for each quantile
\cr (Inherited from \code{\linkS4class{GEVAGroupSet}})}

\item{\code{offsets}}{\code{numeric SVTable} (\emph{m} lines) with the S and V coordinate offsets each gene/probe from its quantile centroid
\cr (Inherited from \code{\linkS4class{GEVAGroupSet}})}

\item{\code{info}}{\code{list} of additional information
\cr (Inherited from \code{\linkS4class{GEVAGroupSet}})}

\item{\code{svscores}}{\code{numeric \linkS4class{SVTable}} (\emph{m} lines) with individual partial scores for the assigned quantiles}

\item{\code{qareasizes}}{\code{numeric \linkS4class{SVTable}} (\emph{g} lines) with the S and V sizes for each quantile}

\item{\code{qindexes}}{\code{integer \linkS4class{SVTable}} (\emph{g} lines) representing the position index to each quantile, in terms of summary and variation}

\item{\code{qcount}}{integer attributes (\code{\linkS4class{SVIntAttribute}}) with the defined number of quantiles for the S and V axes}

\item{\code{qcutoff}}{numeric attributes (\code{\linkS4class{SVNumAttribute}}) with the initial quantile cutoff in S and V, starting from the point zero}
}}

\section{Methods}{
(See also the inherited methods from \code{\linkS4class{GEVAGroupSet}})\cr 

\sspace\cr\strong{Dimension accessors}
\describe{
\item{\code{lines(x, ...)}}{Draws the quantile delimiter lines}
}
\sspace\cr\strong{Slot accessors}
\describe{
\item{\code{qareasizes(object)}}{Gets the \code{SVTable} from the \code{qareasizes} slot}
\item{\code{qcount(object)}}{Gets the \code{SVIntAttribute} from the \code{qcount} slot}
\item{\code{qindexes(object)}}{Gets the \code{SVTable} from the \code{qindexes} slot}
\item{\code{quantiles(object)}}{Gets the unique quantile names}
\item{\code{quantiles.method(object)}}{Gets the \code{character} from the \code{qmethod} slot}
\item{\code{sv.scores(object)}}{Gets the \code{SVTable} from the \code{svscores} slot}
}
}

