% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVAResults.R
\docType{class}
\name{GEVAResults-class}
\alias{GEVAResults-class}
\title{GEVA Results Table}
\description{
The \code{GEVAResults} class contains the final results from GEVA analyses. It represents the results of multiple statistical approaches from summary/variation data, groupings and quantiles.
}
\section{Slots}{

\describe{
\item{\code{resultstable}}{data.frame with classification results for data points}

\item{\code{svdata}}{GEVASummary object with summarized data}

\item{\code{quantdata}}{GEVAQuantiles object with final quantiles definitions (e.g. adjusted quantiles)}

\item{\code{factoring}}{data.frame with detailed factoring information}

\item{\code{classiftable}}{data.frame used as reference to classify the data points}
}}

