% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_GEVASummary.R
\docType{class}
\name{GEVASummary-class}
\alias{GEVASummary-class}
\title{GEVA Summary-Variation Table}
\description{
The \code{GEVASummary} class inherits from \code{SVTable}.
}
\section{Slots}{

\describe{
\item{\code{inputdata}}{GEVAInput-class with the data input}

\item{\code{sv.method}}{Names of the statistical methods used to summarize data}

\item{\code{info}}{list with additional information}
}}

